requirements.txt
Package      Version
------------ -------
grpcio       1.60.0
grpcio-tools 1.60.0
pip          22.0.4
protobuf     4.25.1
setuptools   58.1.0
